Created Genesis Package
Took out what pertained to the Genesis package from Producer and it's tests. Also moved "genesisHashFile" variable to constants
