+++ internal/constants/constants.go
@@ -6,4 +6,5 @@ const (
 	BlockchainFile    = "blockchain.dat"
 	GenesisAddresses  = "genesis_hashes.txt"
 	ConfigurationFile = "config.json"
+	GenesisHashFile   = "genesis_hashes.txt"
 )
